All changes after the reviews
