+++ frontend/jest.config.ts
@@ -5,14 +5,15 @@ export default {
     '!src/**/*.d.ts',
     '!src/components/**',
     '!src/**/index.ts',
-    '!src/helpers/**',
     '!src/main.tsx',
     '!src/reportWebVitals.ts',
     '!src/setupTests.ts',
     '!src/utils/**',
-    '!src/ErrorWrapper.tsx',
     '!src/sentry.config.ts',
     '!src/hooks/**',
+    '!src/wrappers/**',
+    '!src/types/**',
+    '!src/api/**',
   ],
   coverageDirectory: 'coverage',
   coverageReporters: ['text', 'lcov', 'html', 'json-summary'],
