+++ frontend/src/components/Card.tsx
@@ -1,10 +1,11 @@
 import { FontAwesomeIcon, FontAwesomeIconProps } from '@fortawesome/react-fontawesome'
-import { CardProps, tooltipStyle } from 'helpers/constants'
-import FontAwesomeIconWrapper from 'helpers/wrappers/FontAwesomeIconWrapper'
 import { useState, useEffect } from 'react'
 import { Tooltip } from 'react-tooltip'
+import { CardProps } from 'types/card'
+import { tooltipStyle } from 'utils/constants'
 import { Icons } from 'utils/data'
-import { cn } from 'utils/utils'
+import { cn } from 'utils/utility'
+import FontAwesomeIconWrapper from 'wrappers/FontAwesomeIconWrapper'
 import ActionButton from 'components/ActionButton'
 import ContributorAvatar from 'components/ContributorAvatar'
 import DisplayIcon from 'components/DisplayIcon'
