+++ frontend/src/components/ui/Toast.tsx
@@ -3,7 +3,7 @@ import { cva, type VariantProps } from 'class-variance-authority'
 import { X } from 'lucide-react'
 import * as React from 'react'
 
-import { cn } from 'utils/utils'
+import { cn } from 'utils/utility'
 
 const ToastProvider = ToastPrimitives.Provider
