+++ frontend/src/main.tsx
@@ -1,4 +1,3 @@
-import { ErrorWrapper } from 'helpers/wrappers/ErrorWrapper.tsx'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
@@ -6,6 +5,7 @@ import TagManager from 'react-gtm-module'
 import { BrowserRouter } from 'react-router-dom'
 
 import { GTM_AUTH, GTM_ID, GTM_PREVIEW } from 'utils/credentials.ts'
+import { ErrorWrapper } from 'wrappers/ErrorWrapper.tsx'
 import App from './App.tsx'
 
 const tagManagerArgs = {
