+++ frontend/src/pages/ProjectDetails.tsx
@@ -1,9 +1,9 @@
-import { fetchAlgoliaData } from 'helpers/api'
-import FontAwesomeIconWrapper from 'helpers/wrappers/FontAwesomeIconWrapper'
+import { fetchAlgoliaData } from 'api/fetchAlgoliaData'
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { level } from 'utils/data'
-import { getFilteredIcons } from 'utils/utils'
+import { getFilteredIcons } from 'utils/utility'
+import FontAwesomeIconWrapper from 'wrappers/FontAwesomeIconWrapper'
 import Card from 'components/Card'
 import LoadingSpinner from 'components/LoadingSpinner'
