+++ frontend/src/types/button.ts
@@ -1,0 +1,8 @@
+import { JSX } from 'react'
+
+export type ButtonType = {
+  label: string
+  icon?: JSX.Element
+  onclick?: () => void
+  url?: string
+}
