+++ frontend/src/types/contributor.ts
@@ -1,0 +1,6 @@
+export type topContributorsType = {
+  avatar_url: string
+  contributions_count: number
+  login: string
+  name: string
+}
