+++ frontend/src/types/icon.ts
@@ -1,0 +1,3 @@
+export type IconType = {
+  [key: string]: string | number
+}
