+++ frontend/src/types/level.ts
@@ -1,0 +1,5 @@
+export interface Level {
+  color: string
+  icon: string
+  level?: string
+}
