+++ frontend/src/types/link.ts
@@ -1,0 +1,5 @@
+export interface Link {
+  text: string
+  href: string
+  isSpan?: boolean
+}
