+++ frontend/src/types/section.ts
@@ -1,0 +1,6 @@
+import { Link } from './link'
+
+export interface Section {
+  title: string
+  links: Link[]
+}
