+++ frontend/src/utils/constants.ts
@@ -1,10 +1,5 @@
-import { JSX } from 'react'
-
-export interface Link {
-  text: string
-  href: string
-  isSpan?: boolean
-}
+import { Link } from 'types/link'
+import { Section } from 'types/section'
 
 export const headerLinks: Link[] = [
   {
@@ -24,11 +19,6 @@ export const headerLinks: Link[] = [
     href: '/community/users',
   },
 ]
-
-export interface Section {
-  title: string
-  links: Link[]
-}
 
 export const footerSections: Section[] = [
   {
@@ -82,54 +72,6 @@ export const footerSections: Section[] = [
     ],
   },
 ]
-
-export type IconType = {
-  [key: string]: string | number
-}
-
-export interface Level {
-  color: string
-  icon: string
-  level?: string
-}
-
-export type topContributorsType = {
-  avatar_url: string
-  contributions_count: number
-  login: string
-  name: string
-}
-
-export type ButtonType = {
-  label: string
-  icon?: JSX.Element
-  onclick?: () => void
-  url?: string
-}
-
-export interface UserCardProps {
-  avatar: string
-  name: string
-  company: string
-  button: ButtonType
-}
-
-export interface CardProps {
-  title: string
-  url: string
-  summary: string
-  level?: Level
-  icons?: IconType
-  leaders?: string[]
-  topContributors?: topContributorsType[]
-  topics?: string[]
-  button: ButtonType
-  projectName?: string
-  projectLink?: string
-  languages?: string[]
-  social?: { title: string; icon: string; url: string }[]
-  tooltipLabel?: string
-}
 
 export const tooltipStyle = {
   borderRadius: '8px',
