+++ frontend/src/utils/utility.ts
@@ -1,10 +1,10 @@
 import { type ClassValue, clsx } from 'clsx'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { IconType } from 'helpers/constants'
 import { twMerge } from 'tailwind-merge'
 
 import { CommitteeType } from 'types/committee'
+import { IconType } from 'types/icon'
 import { IssueType } from 'types/issue'
 import { project } from 'types/project'
 import { IconKeys, Icons, urlMappings } from 'utils/data'
