+++ frontend/src/wrappers/FontAwesomeIconWrapper.tsx
